endpoint-client/requirements.txt
google-auth
google-cloud
google-cloud-aiplatform
flask
flask-cors
Werkzeug
